// asset-management/src/generated/models/Cr240_assetcategoriesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr240_assetcategoriesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr240_assetcategoriesstatecode = keyof typeof Cr240_assetcategoriesstatecode;
export const Cr240_assetcategoriesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr240_assetcategoriesstatuscode = keyof typeof Cr240_assetcategoriesstatuscode;

export interface Cr240_assetcategoriesBase {
  cr240_assetcategoryid: string;
  cr240_assetcategoryname?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Cr240_assetcategoriesstatecode;
  statuscode?: Cr240_assetcategoriesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr240_assetcategories extends Cr240_assetcategoriesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}